Revert "Merge pull request #7 from Jimcl6/hardening/phase-4-test-ci"
This reverts commit 2c63b4686c79f8178e35d40e73097a6e002895be, reversing
changes made to 944fe443cd2de42b709caeef4f1c9007a856e50b.

+++ app/Imports/TorqueChecksheetImport.php
@@ -522,7 +522,7 @@ class TorqueChecksheetImport
 
         // Handle Excel numeric time (0.0 - 1.0)
         if (is_numeric($timeStr)) {
-            $totalMinutes = (int) round((float) $timeStr * 24 * 60);
+            $totalMinutes = round((float) $timeStr * 24 * 60);
             $hours = intdiv($totalMinutes, 60);
             $minutes = $totalMinutes % 60;
